+++ lib/pages/settings/settings_card.dart
@@ -19,13 +19,14 @@ class SettingsCard extends StatelessWidget {
     void onTap() => beamToNamed(path);
 
     return Card(
-      color: colorConfig().entryCardColor,
-      elevation: 8,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(8),
-      ),
+      //color: colorConfig().entryCardColor,
+      color: Colors.white,
+      elevation: 0,
       child: ListTile(
-        contentPadding: const EdgeInsets.symmetric(horizontal: 32, vertical: 8),
+        contentPadding: const EdgeInsets.symmetric(
+          horizontal: 32,
+          vertical: 8,
+        ),
         leading: icon,
         title: Text(
           title,
@@ -33,6 +34,20 @@ class SettingsCard extends StatelessWidget {
         ),
         onTap: onTap,
       ),
+    );
+  }
+}
+
+class SettingsDivider extends StatelessWidget {
+  const SettingsDivider({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Divider(
+      color: Colors.black,
+      thickness: 1,
+      indent: 24,
+      endIndent: 24,
     );
   }
 }