+++ lib/pages/settings/settings_page.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:lotti/pages/settings/settings_card.dart';
-import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/app_bar/app_bar_version.dart';
 
 class SettingsPage extends StatefulWidget {
@@ -27,12 +26,12 @@ class _SettingsPageState extends State<SettingsPage> {
     final localizations = AppLocalizations.of(context)!;
 
     return Scaffold(
-      backgroundColor: colorConfig().bodyBgColor,
+      //backgroundColor: colorConfig().bodyBgColor,
+      backgroundColor: Colors.white,
       appBar: VersionAppBar(title: localizations.navTabTitleSettings),
       body: Container(
         margin: const EdgeInsets.symmetric(
           vertical: 8,
-          horizontal: 4,
         ),
         child: ListView(
           children: [
@@ -40,22 +39,27 @@ class _SettingsPageState extends State<SettingsPage> {
               title: localizations.settingsTagsTitle,
               path: '/settings/tags',
             ),
+            const SettingsDivider(),
             SettingsCard(
               title: localizations.settingsDashboardsTitle,
               path: '/settings/dashboards',
             ),
+            const SettingsDivider(),
             SettingsCard(
               title: localizations.settingsMeasurablesTitle,
               path: '/settings/measurables',
             ),
+            const SettingsDivider(),
             SettingsCard(
               title: localizations.settingsHealthImportTitle,
               path: '/settings/health_import',
             ),
+            const SettingsDivider(),
             SettingsCard(
               title: localizations.settingsFlagsTitle,
               path: '/settings/flags',
             ),
+            const SettingsDivider(),
             SettingsCard(
               title: localizations.settingsAdvancedTitle,
               path: '/settings/advanced',